requirements.txt
Flask>=3.1,<4.0
requests>=2.32,<3.0
biopython>=1.85,<2.0
python-docx>=1.1,<2.0
matplotlib>=3.8,<4.0
reportlab>=4.0,<5.0
